newbackend/requirement.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==4.1.0
blinker==1.6.2
Brotli==1.0.9
cachelib==0.9.0
celery==5.3.1
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.0
cssselect2==0.7.0
cycler==0.11.0
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
flask_weasyprint==1.0.0
fonttools==4.42.0
greenlet==2.0.1
html5lib==1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.3.1
MarkupSafe==2.1.1
matplotlib==3.7.2
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
pycparser==2.21
pydyf==0.7.0
PyJWT==2.6.0
pyparsing==3.0.9
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.6.0
six==1.16.0
SQLAlchemy==1.4.45
tinycss2==1.2.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.2.2
zopfli==0.2.2